move everything from data.properties back into data - v3.0 port
